refactor: relocated ServicePhoneNumber from Emails to Services namespace
