+++ tests/Sinch.Tests/Fax/EmailsTests.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using RichardSzalay.MockHttp;
 using Sinch.Fax.Emails;
+using Sinch.Fax.Services;
 using Xunit;
 
 namespace Sinch.Tests.Fax
